Update README.md
Predicting trends in stock market prices has been an area of interest for researchers for many years due to its complex and dynamic nature. It is a very challenging task due to the many uncertainties involved and many variables that influence the market value in a particular day such as economic conditions, investors’ sentiments towards a particular company, political events etc. Because of this, stock markets are susceptible to quick changes, causing random fluctuations in the stock price. Stock market series are generally dynamic, non-parametric, chaotic and noisy in nature and hence, stock market price movement is considered to be a random process with fluctuations which are more pronounced for short time windows. Market risk, strongly correlated with forecasting errors, needs to be minimized to ensure minimal risk in investment. Thus, I tried to minimize forecasting error by treating the forecasting problem as a classification problem, a popular suite of algorithms in Machine learning.  Here, I tried to propose a novel way to minimize the risk of investment in stock market by predicting the returns of a stock using a class of powerful machine learning algorithms known as ensemble learning. Some of the technical indicators such as Relative Strength Index (RSI), stochastic oscillator etc are used as inputs to train our model. The learning model used is an ensemble of multiple decision trees.